Actors are moving, but it's pretty buggy.
For one, I'm not doing any bounds checking, so we go off the screen.
Also, for the player, after they've first pressed a direction, they continue to
move even if they aren't pressing a key.
Finally, things are moving really fast even though we're modifying the movement
with the elapsed time.
But I'll fix this tomorrow.