Enhancements to dialogue box creation.
Dialogue can now be entered into the XML file such that each tag is an entire
"line" delivered by a particular speaker.  DialogueView parses this such that
the proper number of dialogue boxes are created with an appropriate amount of
left-justified text.  Additionally, the player now has to click withing the
bounds of the SFML window in order to see the text advance.