Implement medium-sized map/pause screen.
The minimap now only shows approximate locations, but nothing about
the terrain.  This was done due to the fact that the scale of the
minimap was so much smaller than the scale of the Player's represen-
tation on it that it didn't accurately represent their location.
It still isn't perfect on the medium map, but it's better than it
was.  You can access it by pressing the `M` key during a level.