DialogueView adapts to newlines in XML.
Specifically this is done to accommodate a specific presentation of Phil's
dialogue.  The change essentially checks for a manual newline, and if it finds
one it updates the running calculations of how much dialogue box real estate we
have left as if we actual ran out of space on that line.