Configuration changes for Level #5.
Removed NPCs and added a bunch of rocks instead, with the damage flag set to
True.  I didn't think it made sense story-wise for the NPCs to be in Homer's
cave: Phil and Diana are the only one's that really like him, so why would they
have reason to go there?  The number of rocks generated requires a bit of
finesse on the part of the player to move through the level safely, but this is
still possible, so I feel that balancing is maintained, plus it's another way to
mix things up a bit.